requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
psycopg2==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing-extensions==4.12.2
uvicorn==0.30.6
